feat: enhance word selection and ranking logic, add closest words feature
- Introduced EXCLUDED_TARGET_WORDS to filter out non-target words during selection.
- Updated compute_rankings to use uint32 for ranks.
- Modified select_target_words to ensure only base forms are selected as targets.
- Added functionality to retrieve and display closest words in the frontend.
- Replaced WinDialog with GameResultCard for better result presentation.
- Implemented streak tracking for daily game participation.
- Removed GiveUpResultDialog and WinDialog components to streamline the UI.
- Updated ShareButton to handle sharing results for both wins and give-ups.

+++ backend/test_prepare_integration.py
@@ -67,7 +67,7 @@ def test_run_pipeline_game_ranks_are_valid():
         n_words = len(vocab)
         data = np.load(os.path.join(tmpdir, "games", "0001.npz"))
         ranks = data["ranks"]
-        assert ranks.dtype == np.uint16
+        assert ranks.dtype == np.uint32
         assert len(ranks) == n_words
         assert ranks.min() == 1
         assert ranks.max() == n_words
